+++ Tests/SwiftLintFrameworkTests/ConfigurationTests.swift
@@ -190,6 +190,40 @@ class ConfigurationTests: XCTestCase {
         )
     }
 
+    func testIncludedExcludedRelativeLocationLevel1() {
+        FileManager.default.changeCurrentDirectoryPath(Mock.Dir.level1)
+
+        // The included path "File.swift" should be put relative to the configuration file
+        // (~> Resources/ProjectMock/File.swift) and not relative to the path where
+        // SwiftLint is run from (~> Resources/ProjectMock/Level1/File.swift)
+        let configuration = Configuration(configurationFiles: ["../custom_included_excluded.yml"])
+        let actualIncludedPath = configuration.includedPaths.first!.bridge()
+            .absolutePathRepresentation(rootDirectory: configuration.rootDirectory)
+        let desiredIncludedPath = "File1.swift".absolutePathRepresentation(rootDirectory: Mock.Dir.level0)
+        let actualExcludedPath = configuration.excludedPaths.first!.bridge()
+            .absolutePathRepresentation(rootDirectory: configuration.rootDirectory)
+        let desiredExcludedPath = "File2.swift".absolutePathRepresentation(rootDirectory: Mock.Dir.level0)
+
+        XCTAssertEqual(actualIncludedPath, desiredIncludedPath)
+        XCTAssertEqual(actualExcludedPath, desiredExcludedPath)
+    }
+
+    func testIncludedExcludedRelativeLocationLevel0() {
+        // Same as testIncludedPathRelatedToConfigurationFileLocationLevel1(),
+        // but run from the directory the config file resides in
+        FileManager.default.changeCurrentDirectoryPath(Mock.Dir.level0)
+        let configuration = Configuration(configurationFiles: ["custom_included_excluded.yml"])
+        let actualIncludedPath = configuration.includedPaths.first!.bridge()
+            .absolutePathRepresentation(rootDirectory: configuration.rootDirectory)
+        let desiredIncludedPath = "File1.swift".absolutePathRepresentation(rootDirectory: Mock.Dir.level0)
+        let actualExcludedPath = configuration.excludedPaths.first!.bridge()
+            .absolutePathRepresentation(rootDirectory: configuration.rootDirectory)
+        let desiredExcludedPath = "File2.swift".absolutePathRepresentation(rootDirectory: Mock.Dir.level0)
+
+        XCTAssertEqual(actualIncludedPath, desiredIncludedPath)
+        XCTAssertEqual(actualExcludedPath, desiredExcludedPath)
+    }
+
     private class TestFileManager: LintableFileManager {
         func filesToLint(inPath path: String, rootDirectory: String? = nil) -> [String] {
             switch path {
@@ -281,7 +315,7 @@ class ConfigurationTests: XCTestCase {
     func testCustomConfiguration() {
         let file = SwiftLintFile(path: Mock.Swift._0)!
         XCTAssertNotEqual(Mock.Config._0.configuration(for: file),
-                          Mock.Config._0CustomPath.configuration(for: file))
+                          Mock.Config._0Custom.configuration(for: file))
     }
 
     func testConfigurationWithSwiftFileAsRoot() {
@@ -292,7 +326,7 @@ class ConfigurationTests: XCTestCase {
     }
 
     func testConfigurationWithSwiftFileAsRootAndCustomConfiguration() {
-        let configuration = Configuration(configurationFiles: [Mock.Yml._0CustomPath])
+        let configuration = Mock.Config._0Custom
 
         let file = SwiftLintFile(path: Mock.Swift._0)!
         XCTAssertEqual(configuration.configuration(for: file), configuration)